% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_info.R
\name{water_density_at_ambient_temperature}
\alias{water_density_at_ambient_temperature}
\title{Water Info}
\usage{
water_density_at_ambient_temperature(ambient_temp)
}
\arguments{
\item{ambient_temp}{this parameter serves to test whether the user knows ambient temperature to be around 293 K.}
}
\value{
The density of water in kg/L at ambient temperature.
}
\description{
H2O is a universal solvent with vast uses. This package contains a data point specifying the density of water at room temperature.
}
\examples{
temp <- 293
rho <- water_density_at_ambient_temperature(temp)
}
